+++ lib/vtls/vtls.c
@@ -941,7 +941,7 @@ static CURLcode pubkey_pem_to_der(const char *pem,
   if(!begin_pos)
     return CURLE_BAD_CONTENT_ENCODING;
 
-  pem_count = begin_pos - pem;
+  pem_count = (size_t)(begin_pos - pem);
   /* Invalid if not at beginning AND not directly following \n */
   if(0 != pem_count && '\n' != pem[pem_count - 1])
     return CURLE_BAD_CONTENT_ENCODING;
@@ -954,7 +954,7 @@ static CURLcode pubkey_pem_to_der(const char *pem,
   if(!end_pos)
     return CURLE_BAD_CONTENT_ENCODING;
 
-  pem_len = end_pos - pem;
+  pem_len = (size_t)(end_pos - pem);
 
   stripped_pem = malloc(pem_len - pem_count + 1);
   if(!stripped_pem)
@@ -1412,12 +1412,13 @@ static size_t multissl_version(char *buffer, size_t size)
       bool paren = (selected != available_backends[i]);
 
       if(available_backends[i]->version(vb, sizeof(vb))) {
-        p += msnprintf(p, end - p, "%s%s%s%s", (p != backends ? " " : ""),
+        p += msnprintf(p, (size_t)(end - p), "%s%s%s%s",
+                       (p != backends ? " " : ""),
                        (paren ? "(" : ""), vb, (paren ? ")" : ""));
       }
     }
 
-    backends_len = p - backends;
+    backends_len = (size_t)(p - backends);
   }
 
   if(size) {
